% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_difficulty.R
\name{get_item_difficulty}
\alias{get_item_difficulty}
\title{get_item_difficulty}
\usage{
get_item_difficulty(data, correct, vignettes)
}
\arguments{
\item{data}{A dataframe}

\item{correct}{A string indicating the column name storing if the symptom checker solved the case (TRUE or FALSE)}

\item{vignettes}{A string indicating the column name storing the vignette or vignette number}
}
\value{
A data frame object containing the item difficulty for each vignette
}
\description{
Calculates the item difficulty for each case / vignette
}
\examples{
data(symptomcheckRdata)
item_difficulty <- get_item_difficulty(
  data = symptomcheckRdata,
  correct = "Correct_Triage_Advice_provided_from_app",
  vignettes = "Vignette_id"
  )
}
