% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_accuracy.R
\name{plot_accuracy}
\alias{plot_accuracy}
\title{plot_accuracy}
\usage{
plot_accuracy(input)
}
\arguments{
\item{input}{A dataframe containing the output of get_accuracy()}
}
\value{
A ggplot object visualizing the accuracy for either one or multiple symptom checkers
}
\description{
Plots the accuracy for one or multiple symptom checkers
}
\examples{
data(symptomcheckRdata)
accuracy <- get_accuracy(
  data = symptomcheckRdata,
  correct = "Correct_Triage_Advice_provided_from_app",
  apps = "App_name"
  )
accuracy_plot <- plot_accuracy(accuracy)
}
