% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_accuracy_by_triage.R
\name{plot_accuracy_by_triage}
\alias{plot_accuracy_by_triage}
\title{plot_accuracy_by_triage}
\usage{
plot_accuracy_by_triage(input)
}
\arguments{
\item{input}{A dataframe containing the output of get_accuracy_by_triage()}
}
\value{
A ggplot object visualizing the accuracy on each triage level for either one or multiple symptom checkers
}
\description{
Plots the accuracy on each triage level for one or multiple symptom checkers
}
\examples{
data(symptomcheckRdata)
accuracy_by_triage <- get_accuracy_by_triage(
  data = symptomcheckRdata,
  correct = "Correct_Triage_Advice_provided_from_app",
  triagelevel = "Goldstandard_solution",
  apps = "App_name"
  )
accuracy_triage_plot <- plot_accuracy_by_triage(accuracy_by_triage)
}
