% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ccs.R
\name{plot_ccs}
\alias{plot_ccs}
\title{plot_ccs}
\usage{
plot_ccs(input)
}
\arguments{
\item{input}{A dataframe containing the output of get_ccs()}
}
\value{
A ggplot object visualizing the CCS for multiple symptom checkers
}
\description{
Plots the Capability Comparison Score (CCS) for multiple sympotm checkers
}
\examples{
data(symptomcheckRdata)
ccs <- get_ccs(
  data = symptomcheckRdata,
  correct = "Correct_Triage_Advice_provided_from_app",
  vignettes = "Vignette_id",
  apps = "App_name"
  )
ccs_plot <- plot_ccs(ccs)
}
