% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inclination_to_overtriage.R
\name{plot_inclination_to_overtriage}
\alias{plot_inclination_to_overtriage}
\title{plot_inclination_to_overtriage}
\usage{
plot_inclination_to_overtriage(input)
}
\arguments{
\item{input}{A dataframe containing the output of get_inclination_to_overtriage()}
}
\value{
A ggplot object visualizing the inclination to overtriage for either one or multiple symptom checkers
}
\description{
Plots the inclination to overtriage for one or multiple symptom checkers
}
\examples{
data(symptomcheckRdata)
inclination_to_overtriage <- get_inclination_to_overtriage(
  data = symptomcheckRdata,
  triagelevel_correct = "Goldstandard_solution",
  triagelevel_advice = "Triage_advice_from_app",
  order_triagelevel = c("Emergency", "Non-Emergency", "Self-care"),
  apps = "App_name"
  )
overtriage_plot <- plot_inclination_to_overtriage(inclination_to_overtriage)
}
